refactor: combine login and register into shared layout
