+++ src/features/auth/__tests__/LoginComponent.test.tsx
@@ -1,0 +1,27 @@
+// import { render, screen } from '@testing-library/react';
+// import { AuthLayout } from '../components/auth-layout/AuthLayout';
+
+// describe('LoginComponent', () => {
+//   test('should show heading', () => {
+//     render(<AuthLayout />);
+//     expect(
+//       screen.getByRole('heading', { name: /sign in/i }),
+//     ).toBeInTheDocument();
+//   });
+//   test('should show paragraph text', () => {
+//     render(<AuthLayout />);
+
+//     expect(
+//       screen.getByText(/Welcome back! Please enter your details/i),
+//     ).toBeInTheDocument();
+//   });
+//   test('should render img', () => {
+//     render(<AuthLayout />);
+
+//     expect(screen.getByAltText('signin-img')).toBeInTheDocument();
+//   });
+//   test('should render form', () => {
+//     render(<AuthLayout />);
+//     expect(screen.getByTestId('signin-form')).toBeInTheDocument();
+//   });
+// });
