+++ src/features/auth/components/auth-layout/AuthLayout.module.css
@@ -2,39 +2,37 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  height: auto;
 }
 .content {
   display: flex;
-  /* height: 850px; */
   height: auto;
   width: fit-content;
+  border-radius: var(--border-radius-block);
+}
+.container {
+  padding: 2rem 0;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 550px;
+  height: 100%;
+  color: var(--color-primary-dark);
+  background-color: var(--color-background-ui);
   border-radius: var(--border-radius-block);
 }
 .logo {
   font-size: 32px;
 }
-.container {
-  padding: 1rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 650px;
-  height: 100%;
-  color: var(--color-primary-dark);
-  background-color: var(--color-background-ui);
-  border-radius: var(--border-radius-block) 0 0 var(--border-radius-block);
-}
 .container > h2 {
-  color: var(--color-primary-light);
+  color: var(--color-primary-dark);
   font-size: 32px;
-  margin-top: 1rem;
+  margin-top: 0.75rem;
 }
 .container > p {
   font-size: 16px;
   margin-top: 10px;
-  margin-bottom: 50px;
+  margin-bottom: 30px;
 }
 
 @media (max-width: 1024px) {
