--- src/features/auth/login/components/form/LoginForm.module.css
@@ -1,35 +1,0 @@
-.form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: auto;
-  gap: 20px;
-}
-
-.button {
-  width: 250px;
-  height: 50px;
-  margin-top: 20px;
-  cursor: pointer;
-}
-.buttons {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-.buttons > .button:nth-child(2) {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-  background-color: var(--color-background-ui);
-  color: var(--color-primary-dark);
-  border: 1px solid var(--color-primary-dark);
-  font-size: 16px;
-  margin-bottom: 20px;
-}
-.signup {
-  color: var(--color-primary-light);
-}
