--- src/features/auth/login/components/login-container/LoginContainer.module.css
@@ -1,50 +1,0 @@
-.main {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.content {
-  display: flex;
-  height: auto;
-  width: fit-content;
-  border-radius: var(--border-radius-block);
-}
-.container {
-  padding: 2rem 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 550px;
-  height: 100%;
-  color: var(--color-primary-dark);
-  background-color: var(--color-background-ui);
-  border-radius: var(--border-radius-block) 0 0 var(--border-radius-block);
-}
-.logo {
-  font-size: 32px;
-}
-.container > h2 {
-  color: var(--color-primary-dark);
-  font-size: 32px;
-  margin-top: 1rem;
-}
-.container > p {
-  font-size: 16px;
-  margin-top: 10px;
-  margin-bottom: 50px;
-}
-
-@media (max-width: 1024px) {
-  .container {
-    margin: 0 auto;
-  }
-  .content {
-    width: 95%;
-  }
-}
-@media (max-width: 425px) {
-  .container > p {
-    font-size: 12px;
-  }
-}
