--- src/features/auth/login/styles/LoginPage.module.css
@@ -1,6 +1,0 @@
-.content {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh;
-}
