+++ src/pages/LoginPage.tsx
@@ -1,11 +1,18 @@
-import { LoginContainer } from '../features/auth/login/components/login-container/LoginContainer';
-import styles from '../features/auth/login/styles/LoginPage.module.css';
+import { AuthLayout } from '../features/auth/components/auth-layout/AuthLayout';
+import { LoginForm } from '../features/auth/components/form/LoginForm';
+import styles from '../features/auth/styles/AuthPages.module.css';
 export const LoginPage = () => {
+  const text = 'Welcome back! Please enter your details';
+  const title = 'Sign in';
   return (
-    <div className="container">
-      <main className={styles.content}>
-        <LoginContainer />
-      </main>
+    <div className={styles.auth}>
+      <div className="container">
+        <main className={styles.content}>
+          <AuthLayout title={title} text={text}>
+            <LoginForm />
+          </AuthLayout>
+        </main>
+      </div>
     </div>
   );
 };
